Refactor: IronField to module
